[CMPT-5391] extract shared HandleRunNotFoundError into runutil, remove duplicate handlers
