+++ cmd/pipeline/run/cancel/cmd.go
@@ -17,11 +17,10 @@ package cancel
 import (
 	"errors"
 	"fmt"
-	"net/http"
 
+	"github.com/datarobot/cli/cmd/pipeline/run/runutil"
 	"github.com/datarobot/cli/cmd/pipeline/scopeflag"
 	"github.com/datarobot/cli/internal/auth"
-	"github.com/datarobot/cli/internal/drapi"
 	"github.com/datarobot/cli/internal/pipeline"
 	"github.com/datarobot/cli/internal/telemetry"
 	"github.com/datarobot/cli/tui"
@@ -57,7 +56,7 @@ Example:
 
 			err = pipeline.CancelRun(flags.PipelineID, scope, version, args[0])
 			if err != nil {
-				return handleCancelError(err, args[0])
+				return runutil.HandleRunNotFoundError(err, args[0])
 			}
 
 			fmt.Println(tui.BaseTextStyle.Render("Cancelled run: " + args[0]))
@@ -79,16 +78,4 @@ Example:
 	})
 
 	return cmd
-}
-
-func handleCancelError(err error, runID string) error {
-	var httpErr *drapi.HTTPError
-
-	if errors.As(err, &httpErr) && httpErr.StatusCode == http.StatusNotFound {
-		fmt.Println(tui.DimStyle.Render("No run found with id: " + runID))
-
-		return nil
-	}
-
-	return err
 }
